fix DWORD issue in header file

+++ src/inference/dev_api/ie_system_conf.h
@@ -191,7 +191,7 @@ typedef enum {
  */
 INFERENCE_ENGINE_API_CPP(void)
 parse_processor_info_win(char* base_ptr,
-                         DWORD len,
+                         unsigned long len,
                          int& _processors,
                          int& _sockets,
                          int& _cores,
