+++ src/inference/src/os/win/win_system_conf.cpp
@@ -43,7 +43,7 @@ struct CPU {
 static CPU cpu;
 
 void parse_processor_info_win(char* base_ptr,
-                              DWORD len,
+                              unsigned long len,
                               int& _processors,
                               int& _sockets,
                               int& _cores,
